backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
bosdyn-client>=4.0.0
bosdyn-api>=4.0.0
bosdyn-core>=4.0.0
python-dotenv>=1.0.0
